Fix printf format warnings wrt int/long size.

+++ events.c
@@ -47,7 +47,7 @@ static void list_events(const char *arg, struct session *ses)
         tintin_printf(ses, "#Defined events:");
         while (ev)
         {
-            tintin_printf(ses, "(%d.%03d)\t {%s}", (int)((ev->time-ct)/NANO),
+            tintin_printf(ses, "(%lld.%03d)\t {%s}", (ev->time-ct)/NANO,
                 msec(ev->time-ct), ev->event);
             ev = ev->next;
         }
@@ -59,7 +59,7 @@ static void list_events(const char *arg, struct session *ses)
         {
             if (match(left, ev->event))
             {
-                tintin_printf(ses, "(%d.%03d)\t {%s}", (ev->time-ct)/NANO,
+                tintin_printf(ses, "(%lld.%03d)\t {%s}", (ev->time-ct)/NANO,
                     msec(ev->time-ct), ev->event);
                 flag = true;
             }
@@ -162,7 +162,7 @@ void undelay_command(const char *arg, struct session *ses)
         {
             flag=true;
             if (ses==activesession && ses->mesvar[MSG_EVENT])
-                tintin_printf(ses, "#Ok. Event {%s} at %ld.%03d won't be executed.",
+                tintin_printf(ses, "#Ok. Event {%s} at %lld.%03d won't be executed.",
                     (*ev)->event, ((*ev)->time-ct)/NANO, msec((*ev)->time-ct));
             remove_event(ev);
         }
